+++ routes/registerRoutes/patientRegisterRoute.js
@@ -1,7 +1,8 @@
 const express = require('express');
-const orm_registerPatient = require('../../ORM/ORM_registerPatient');
+// const orm_registerPatient = require('../../ORM/ORM_registerPatient');
 const cookie = require('../../authentication/cookie_checker');
 require('dotenv').config();
+const {patientRegisterController} = require('../../controllers/registercontroller'); // Import the patientRegisterController function
 
 
 const router = express.Router({mergeParams : true});
@@ -20,27 +21,7 @@ router.post('/api/v0/register_patient', async (req, res) => { //post->get
     }
 });
 
-async function patientRegisterController(req) {
-    const req_json = req.body;
 
-    // if(!cookie.checkCookie(req.session)) {
-    //     // Force to login. And return a status
-    //     return;
-    // }
-
-    console.log(req_json);
-    const model_patient_basic_info = await orm_registerPatient.insertPatientInfo(req_json);
-    if(model_patient_basic_info) {
-        console.log('Patient Registered!');
-        // console.log(model_patient_basic_info);
-        const pseudo_view = model_patient_basic_info;
-        return pseudo_view;
-    }
-    else {
-        console.log('NO user Found!');
-        return null;
-    }
-}
 
 
 module.exports = router;
