+++ routes/searchRoutes/searchPatientRoute.js
@@ -1,5 +1,6 @@
 const express = require('express');
-const orm_search_patient = require('../../ORM/ORM_patient_info');
+// const orm_search_patient = require('../../ORM/ORM_patient_info');
+const { patientSearchController } = require('../../controllers/searchcontroller'); // Import the patientSearchController function
 const cookie = require('../../authentication/cookie_checker');
 require('dotenv').config();
 
@@ -21,26 +22,7 @@ router.post('/api/v0/search_patient', async (req, res) => { //post->get
     }
 });
 
-async function patientSearchController(req) {
-    const req_json = req.body;
 
-    // if(!cookie.checkCookie(req.session)) {
-    //     // Force to login. And return a status
-    //     return;
-    // }
-    console.log(req_json);
-    const model_patient_basic_info = await orm_search_patient.get_patient_basic_info(req_json.phone);
-    if(model_patient_basic_info) {
-        console.log('Patient found!');
-        console.log(model_patient_basic_info);
-        const pseudo_view = model_patient_basic_info;
-        return pseudo_view;
-    }
-    else {
-        console.log('NO user Found!');
-        return null;
-    }
-}
